so many changes ...

+++ app/Helpers/Booking/BookingService.php
@@ -3,17 +3,11 @@
 namespace App\Helpers\Booking;
 
 use App\Models\Booking as BookingModel;
-use App\Models\Manager;
-use App\Models\Passenger;
 use App\Models\Payment;
 use App\Models\Room;
 use App\Models\Unbookable;
-use App\Models\User;
 use Carbon\Carbon;
-use Illuminate\Support\Facades\Log;
 use illuminate\Support\Str;
-use App\Helpers\Logs\Logs;
-
 use Shetabit\Multipay\Invoice;
 use Shetabit\Payment\Facade\Payment as ShetabitPayment;
 use Shetabit\Multipay\Exceptions\InvalidPaymentException;
@@ -26,31 +20,19 @@ class BookingService {
     public function putBookingToSession()
     {
         $id = Str::random(40);
-        $user = auth('web')->user();
-        $room = Booking::getRoom();
-        $checkin = Booking::getCheckin();
-        $checkinJalali = Booking::getCheckinJalali();
-        $checkout = Booking::getCheckout();
-        $checkoutJalali = Booking::getCheckoutJalali();
-        $length = Booking::getLength();
-        $adults = Booking::getAdults();
-        $teacher = Booking::getTeacher();
-        $passengers = Booking::getPassengers();
 
-
-        
         session()->push('bookings', collect([
                 'id' => $id,
-                'user' => $user,
-                'room' => $room,
-                'checkin' => $checkin,
-                'checkout' => $checkout,
-                'checkinJalali' => $checkinJalali,
-                'checkoutJalali' => $checkoutJalali,
-                'length' => $length,
-                'adults' => $adults,
-                'teacher' => $teacher,
-                'passengers' => $passengers,
+                'user' => user('web'),
+                'room' => Booking::getRoom(),
+                'checkin' => Booking::getCheckin(),
+                'checkout' => Booking::getCheckout(),
+                'checkinJalali' => Booking::getCheckinJalali(),
+                'checkoutJalali' => Booking::getCheckoutJalali(),
+                'length' => Booking::getLength(),
+                'adults' => Booking::getAdults(),
+                'teacher' => Booking::getTeacher(),
+                'passengers' => Booking::getPassengers(),
             ])
         );
 
@@ -78,7 +60,7 @@ class BookingService {
     public function newBooking($booking)
     {
         $user = $booking->get('user');
-        $room = Room::find($booking->get('room')->id);
+        $room = $booking->get('room');
         $checkin = $booking->get('checkin');
         $checkout = $booking->get('checkout');
         $amount = $booking->get('length') * $room->price;
@@ -93,7 +75,6 @@ class BookingService {
                 'national_code' => $teacher['national_code'],
             ]
         ];
-
         $head = [
             'head' => [
                 'name' => $passengers[1]['first_name'] . ' ' . $passengers[1]['last_name'],
@@ -101,7 +82,6 @@ class BookingService {
                 'phone' => $passengers[1]['phone'],
             ],
         ];
-        
         $i=1;
         foreach ($passengers as $passenger) {
             $_passengers['passengers'][$i] = [
@@ -111,29 +91,30 @@ class BookingService {
             $i++;
         }
 
+
         $detail = array_merge($teacher, $head, $_passengers);
 
-        $existing_booking = BookingModel::where([['user_id', $user->id], ['room_id', $room->id], ['checkin', $checkin], ['checkout', $checkout]])->first();
+
+        $this->room_number = $this->getValidRoomNumbers($room)->first();
+
+        $existing_booking = BookingModel::where([['user_id', $user->id], ['room_id', $room->id], ['checkin', $checkin], ['checkout', $checkout], ['status', 'unpaid']])->whereIn('room_number', $this->getValidRoomNumbers($room))->first();
 
         if (is_null($existing_booking)) {
-
             // create booking
             $booking = BookingModel::create([
                 'user_id' => $user->id,
                 'room_id' => $room->id,
+                'room_number' => $this->room_number,
                 'checkin' => $checkin,
                 'checkout' => $checkout,
                 'amount' => $amount,
-                'rooms' => $room->numbers,
                 'status' => 'unpaid',
             ]);
-        
 
             // add passengers
             $booking->passengers()->create([
                 'detail' => $detail
             ]);
-
 
             return $this->booking = $booking;
         }
@@ -168,6 +149,7 @@ class BookingService {
             Unbookable::create([
                 'user_id' => $user_id,
                 'room_id' => $room_id,
+                'room_number' => $this->room_number,
                 'start_date' => $start_date,
                 'end_date' => $end_date,
                 'expiration' => $expiration,
@@ -179,16 +161,20 @@ class BookingService {
 
     public function defreezeRoom($booking_id)
     {
+
         $booking = BookingModel::findOrFail($booking_id);
 
         $user_id = $booking->user_id;
         $room_id = $booking->room_id;
         $start_date = $booking->checkin;
         $end_date = $booking->checkout;
+        $room_number = $booking->room_number;
 
-        $unbookable = Unbookable::where([['user_id', $user_id], ['room_id', $room_id], ['start_date', $start_date], ['end_date', $end_date]])->firstOrFail();
+        $unbookable = Unbookable::where([['user_id', $user_id], ['room_id', $room_id], ['room_number', $room_number], ['start_date', $start_date], ['end_date', $end_date]])->first();
 
-        $unbookable->delete();
+        if ($unbookable) {
+            $unbookable->delete();
+        }
     }
 
 
@@ -214,14 +200,12 @@ class BookingService {
         // ]);
 
         return ShetabitPayment::callbackUrl(route('reserve.payment.callback'))->purchase($invoice, function($driver, $transactionId) use ($payment_amount, $manager_amount) {
-
             Payment::create([
                 'booking_id' => $this->booking->id,
                 'amount' => $payment_amount,
                 'booking_amount' => $manager_amount,
                 'track_id' => $transactionId,
             ]);
-
         })->pay()->render();
     }
 
@@ -239,25 +223,26 @@ class BookingService {
                 'status' => 1,
             ]);
 
-
-            while(1) {
-                $voucher = rand(11111111, 99999999);
-                if (! BookingModel::where('voucher', $voucher)->exists()){
-                    break;
+            if (is_null($payment->booking->voucher)) {
+                while(1) {
+                    $voucher = rand(11111111, 99999999);
+                    if (! BookingModel::where('voucher', $voucher)->exists()){
+                        break;
+                    }
                 }
+                $payment->booking()->update([
+                    'voucher' => $voucher,
+                    'status' => 'paid',
+                ]);
             }
-            $payment->booking()->update([
-                'voucher' => $voucher,
-                'status' => 'paid',
-            ]);
 
-            
         } catch (InvalidPaymentException $exception) {
 
             $this->defreezeRoom($payment->booking_id); // it is important to be at the top of the code
 
             echo 'پرداخت با خطا مواجه شد. در صورتی که مبلغ از حساب شما برداشت شده حداکثر تا 72 ساعت به حساب شما باز خواهد گشت';
         }
+
     }
 
 }
